# spec/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /api/agents:
    get:
      summary: Get Agents
      operationId: get_agents_api_agents_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AgentResponse'
                type: array
                title: Response Get Agents Api Agents Get
    post:
      summary: Create Agent
      operationId: create_agent_api_agents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/agents/exists:
    get:
      summary: Exists Agent
      operationId: exists_agent_api_agents_exists_get
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
          title: Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: boolean
                title: Response Exists Agent Api Agents Exists Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/agents/{agent_id}:
    put:
      summary: Update Agent
      operationId: update_agent_api_agents__agent_id__put
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Agent
      operationId: delete_agent_api_agents__agent_id__delete
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/todos:
    get:
      summary: Get Todos
      operationId: get_todos_api_todos_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TodoResponse'
                type: array
                title: Response Get Todos Api Todos Get
    post:
      summary: Create Todo
      operationId: create_todo_api_todos_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTodoRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/todos/{todo_id}/events:
    get:
      summary: Todo Events
      operationId: todo_events_api_todos__todo_id__events_get
      parameters:
      - name: todo_id
        in: path
        required: true
        schema:
          type: string
          title: Todo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AgentRequest:
      properties:
        name:
          type: string
          title: Name
        system_prompt:
          type: string
          title: System Prompt
      type: object
      required:
      - name
      - system_prompt
      title: AgentRequest
    AgentResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        system_prompt:
          type: string
          title: System Prompt
      type: object
      required:
      - id
      - name
      - system_prompt
      title: AgentResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PostTodoRequest:
      properties:
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
      type: object
      required:
      - title
      - content
      title: PostTodoRequest
    TodoResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        status:
          type: string
          title: Status
        assigned_agent_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned Agent Name
      type: object
      required:
      - id
      - title
      - content
      - status
      title: TodoResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError